restructure project and timer
