+++ init.d/02.frontend/escape-room-frontend/src/components/GameStatus.tsx
@@ -1,0 +1,13 @@
+import { useContext } from "react"
+import { GameContext } from "../main"
+
+function GameStatus() {
+    const startTime = useContext(GameContext)
+    return (
+        <>
+            { new Date(new Date().getTime() - startTime.getTime()).toISOString().substring(11, 19) }
+        </>
+    )
+}
+
+export default GameStatus
