+++ init.d/02.frontend/escape-room-frontend/src/main.tsx
@@ -1,13 +1,17 @@
-import { StrictMode } from 'react'
+import { createContext, StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter as Router } from 'react-router-dom';
 import './index.css'
-import App from './App.tsx'
+import App from './views/App.tsx'
+
+export const GameContext = createContext(new Date())
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <Router>
-      <App />
+      <GameContext value={new Date()}>
+        <App />
+      </GameContext>
     </Router>
   </StrictMode>
 )
