+++ init.d/02.frontend/escape-room-frontend/src/views/App.css
